snapcraft: use system certificates by default
The `requests` library honours REQUESTS_CA_BUNDLE=<some-cert>,
but by default relies on `certifi` for the certificates.

If a user installs a certificate, they must also specify
REQUESTS_CA_BUNDLE.  Instead here we change the snapcraft
environment to automatically use the host certs if it exists.
However, honour user-specified REQUESTS_CA_BUNDLE.  This removes
the need to get certifi up-to-date as well.

Note that the `certifi` distro package uses system CA certificates
by default, but the pip-provided package does not.

Signed-off-by: Chris Patterson <chris.patterson@canonical.com>